Webpage Update
The home page is created, providing a description of the website and its
development. Each page is linked to each other page. Still in a rather
unfinished state, but provides a good outline, and leaves most of the
work up to research and typing... and a little bit of Javascript.